Module google_partners2::api

Structs

AdWordsManagerAccountInfo
Information about a particular AdWords Manager Account. Read more at https://support.google.com/adwords/answer/6139186

AnalyticListCall
Lists analytics data for a user's associated company. Should only be called within the context of an authorized logged in user.

AnalyticMethods
A builder providing access to all methods supported on analytic resources. It is not used directly, but through the Partners hub.

Analytics
Analytics data for a Company within a single day.

AnalyticsDataPoint
Details of the analytics events for a Company within a single day.

AnalyticsSummary
Analytics aggregated data for a Company for a given date range.

AvailableOffer
Available Offers to be distributed.

Certification
A user's information on a specific certification.

CertificationExamStatus
Status for a Google Partners certification exam.

CertificationStatus
Google Partners certification status.

ClientMessageLogCall
Logs a generic message from the client, such as Failed to render component, Profile page is running slow, More than 500 users have accessed this result., etc.

ClientMessageMethods
A builder providing access to all methods supported on clientMessage resources. It is not used directly, but through the Partners hub.

Company
A company resource in the Google Partners API. Once certified, it qualifies for being searched by advertisers.

CompanyGetCall
Gets a company.

CompanyLeadCreateCall
Creates an advertiser lead for the given company ID.

CompanyListCall
Lists companies.

CompanyMethods
A builder providing access to all methods supported on company resources. It is not used directly, but through the Partners hub.

CompanyRelation
A CompanyRelation resource representing information about a user's affiliation and standing with a company in Partners.

CountryOfferInfo
Offer info by country.

CreateLeadRequest
Request message for CreateLead.

CreateLeadResponse
Response message for CreateLead.

Date
Represents a whole or partial calendar date, e.g. a birthday. The time of day and time zone are either specified elsewhere or are not significant. The date is relative to the Proleptic Gregorian Calendar. This can represent:

DebugInfo
Debug information about this request.

Empty
A generic empty message that you can re-use to avoid defining duplicated empty messages in your APIs. A typical example is to use it as the request or the response type of an API method. For instance:

EventData
Key value data pair for an event.

ExamStatus
A user's information on a specific exam.

GetCompanyResponse
Response message for GetCompany.

GetPartnersStatusResponse
Response message for GetPartnersStatus.

HistoricalOffer
Historical information about a Google Partners Offer.

LatLng
An object representing a latitude/longitude pair. This is expressed as a pair of doubles representing degrees latitude and degrees longitude. Unless specified otherwise, this must conform to the WGS84 standard. Values must be within normalized ranges.

Lead
A lead resource that represents an advertiser contact for a Company. These are usually generated via Google Partner Search (the advertiser portal).

LeadListCall
Lists advertiser leads for a user's associated company. Should only be called within the context of an authorized logged in user.

LeadMethods
A builder providing access to all methods supported on lead resources. It is not used directly, but through the Partners hub.

ListAnalyticsResponse
Response message for ListAnalytics.

ListCompaniesResponse
Response message for ListCompanies.

ListLeadsResponse
Response message for ListLeads.

ListOffersHistoryResponse
Response for ListOfferHistory.

ListOffersResponse
Response for ListOffer.

ListUserStatesResponse
Response message for ListUserStates.

LocalizedCompanyInfo
The localized company information.

Location
A location with address and geographic coordinates. May optionally contain a detailed (multi-field) version of the address.

LogMessageRequest
Request message for LogClientMessage.

LogMessageResponse
Response message for LogClientMessage.

LogUserEventRequest
Request message for LogUserEvent.

LogUserEventResponse
Response message for LogUserEvent.

MethodGetPartnersstatuCall
Gets Partners Status of the logged in user's agency. Should only be called if the logged in user is the admin of the agency.

MethodMethods
A builder providing access to all free methods, which are not associated with a particular resource. It is not used directly, but through the Partners hub.

MethodUpdateCompanyCall
Update company. Should only be called within the context of an authorized logged in user.

MethodUpdateLeadCall
Updates the specified lead.

Money
Represents an amount of money with its currency type.

OfferCustomer
Customers qualified for an offer.

OfferHistoryListCall
Lists the Historical Offers for the current user (or user's entire company)

OfferListCall
Lists the Offers available for the current user

OfferMethods
A builder providing access to all methods supported on offer resources. It is not used directly, but through the Partners hub.

OptIns
A set of opt-ins for a user.

Partners
Central instance to access all Partners related resource activities

PublicProfile
Basic information from a public profile.

Rank
Information related to ranking of results.

RecaptchaChallenge
reCaptcha challenge info.

RequestMetadata
Common data that is in each API request.

ResponseMetadata
Common data that is in each API response.

SpecializationStatus
Agency specialization status

TrafficSource
Source of traffic for the current request.

User
A resource representing a user of the Partners platform.

UserCreateCompanyRelationCall
Creates a user's company relation. Affiliates the user to a company.

UserDeleteCompanyRelationCall
Deletes a user's company relation. Unaffiliaites the user from a company.

UserEventLogCall
Logs a user event.

UserEventMethods
A builder providing access to all methods supported on userEvent resources. It is not used directly, but through the Partners hub.

UserGetCall
Gets a user.

UserMethods
A builder providing access to all methods supported on user resources. It is not used directly, but through the Partners hub.

UserOverrides
Values to use instead of the user's respective defaults. These are only honored by whitelisted products.

UserProfile
The profile information of a Partners user.

UserStateListCall
Lists states for current user.

UserStateMethods
A builder providing access to all methods supported on userState resources. It is not used directly, but through the Partners hub.

UserUpdateProfileCall
Updates a user's profile. A user can only update their own profile and should only be called within the context of a logged in user.