Struct google_partners2::CompanyGetCall

pub struct CompanyGetCall<'a, C, A> where C: 'a, A: 'a {
    // some fields omitted
}

Gets a company.

A builder for the get method supported by a company resource. It is not used directly, but through a CompanyMethods instance.

Example

Instantiate a resource method builder

 
// You can configure optional parameters by calling the respective setters at will, and
// execute the final call using `doit()`.
// Values shown here are possibly random and not representative !
let result = hub.companies().get("companyId")
             .view("sit")
             .request_metadata_user_overrides_user_id("Stet")
             .request_metadata_user_overrides_ip_address("sed")
             .request_metadata_traffic_source_traffic_sub_id("et")
             .request_metadata_traffic_source_traffic_source_id("dolores")
             .request_metadata_partners_session_id("kasd")
             .request_metadata_locale("accusam")
             .add_request_metadata_experiment_ids("takimata")
             .order_by("justo")
             .currency_code("amet.")
             .address("erat")
             .doit();

Methods

impl<'a, C, A> CompanyGetCall<'a, C, A> where C: BorrowMut<Client>, A: GetToken

fn doit(self) -> Result<(Response, GetCompanyResponse)>

Perform the operation you have build so far.

fn company_id(self, new_value: &str) -> CompanyGetCall<'a, C, A>

The ID of the company to retrieve.

Sets the company id path property to the given value.

Even though the property as already been set when instantiating this call, we provide this method for API completeness.

fn view(self, new_value: &str) -> CompanyGetCall<'a, C, A>

The view of Company resource to be returned. This must not be COMPANY_VIEW_UNSPECIFIED.

Sets the view query property to the given value.

fn request_metadata_user_overrides_user_id(self, new_value: &str) -> CompanyGetCall<'a, C, A>

Logged-in user ID to impersonate instead of the user's ID.

Sets the request metadata.user overrides.user id query property to the given value.

fn request_metadata_user_overrides_ip_address(self, new_value: &str) -> CompanyGetCall<'a, C, A>

IP address to use instead of the user's geo-located IP address.

Sets the request metadata.user overrides.ip address query property to the given value.

fn request_metadata_traffic_source_traffic_sub_id(self, new_value: &str) -> CompanyGetCall<'a, C, A>

Second level identifier to indicate where the traffic comes from. An identifier has multiple letters created by a team which redirected the traffic to us.

Sets the request metadata.traffic source.traffic sub id query property to the given value.

fn request_metadata_traffic_source_traffic_source_id(self, new_value: &str) -> CompanyGetCall<'a, C, A>

Identifier to indicate where the traffic comes from. An identifier has multiple letters created by a team which redirected the traffic to us.

Sets the request metadata.traffic source.traffic source id query property to the given value.

fn request_metadata_partners_session_id(self, new_value: &str) -> CompanyGetCall<'a, C, A>

Google Partners session ID.

Sets the request metadata.partners session id query property to the given value.

fn request_metadata_locale(self, new_value: &str) -> CompanyGetCall<'a, C, A>

Locale to use for the current request.

Sets the request metadata.locale query property to the given value.

fn add_request_metadata_experiment_ids(self, new_value: &str) -> CompanyGetCall<'a, C, A>

Experiment IDs the current request belongs to.

Append the given value to the request metadata.experiment ids query property. Each appended value will retain its original ordering and be '/'-separated in the URL's parameters.

fn order_by(self, new_value: &str) -> CompanyGetCall<'a, C, A>

How to order addresses within the returned company. Currently, only address and address desc is supported which will sorted by closest to farthest in distance from given address and farthest to closest distance from given address respectively.

Sets the order by query property to the given value.

fn currency_code(self, new_value: &str) -> CompanyGetCall<'a, C, A>

If the company's budget is in a different currency code than this one, then the converted budget is converted to this currency code.

Sets the currency code query property to the given value.

fn address(self, new_value: &str) -> CompanyGetCall<'a, C, A>

The address to use for sorting the company's addresses by proximity. If not given, the geo-located address of the request is used. Used when order_by is set.

Sets the address query property to the given value.

fn delegate(self, new_value: &'a mut Delegate) -> CompanyGetCall<'a, C, A>

The delegate implementation is consulted whenever there is an intermediate result, or if something goes wrong while executing the actual API request.

It should be used to handle progress information, and to implement a certain level of resilience.

Sets the delegate property to the given value.

fn param<T>(self, name: T, value: T) -> CompanyGetCall<'a, C, A> where T: AsRef<str>

Set any additional parameter of the query string used in the request. It should be used to set parameters which are not yet available through their own setters.

Please note that this method must not be used to set any of the known paramters which have their own setter method. If done anyway, the request will fail.

Additional Parameters

bearer_token (query-string) - OAuth bearer token.

pp (query-boolean) - Pretty-print response.

prettyPrint (query-boolean) - Returns response with indentations and line breaks.

upload_protocol (query-string) - Upload protocol for media (e.g. "raw", "multipart").

access_token (query-string) - OAuth access token.

fields (query-string) - Selector specifying which fields to include in a partial response.

quotaUser (query-string) - Available to use for quota purposes for server-side applications. Can be any arbitrary string assigned to a user, but should not exceed 40 characters.

callback (query-string) - JSONP

oauth_token (query-string) - OAuth 2.0 token for the current user.

key (query-string) - API key. Your API key identifies your project and provides you with API access, quota, and reports. Required unless you provide an OAuth 2.0 token.

uploadType (query-string) - Legacy upload protocol for media (e.g. "media", "multipart").

alt (query-string) - Data format for response.

$.xgafv (query-string) - V1 error format.