Struct google_partners2::CompanyListCall

pub struct CompanyListCall<'a, C, A> where C: 'a, A: 'a {
    // some fields omitted
}

Lists companies.

A builder for the list method supported by a company resource. It is not used directly, but through a CompanyMethods instance.

Example

Instantiate a resource method builder

 
// You can configure optional parameters by calling the respective setters at will, and
// execute the final call using `doit()`.
// Values shown here are possibly random and not representative !
let result = hub.companies().list()
             .website_url("sea")
             .view("nonumy")
             .add_services("dolores")
             .request_metadata_user_overrides_user_id("gubergren")
             .request_metadata_user_overrides_ip_address("sadipscing")
             .request_metadata_traffic_source_traffic_sub_id("aliquyam")
             .request_metadata_traffic_source_traffic_source_id("ea")
             .request_metadata_partners_session_id("no")
             .request_metadata_locale("justo")
             .add_request_metadata_experiment_ids("justo")
             .page_token("et")
             .page_size(-17)
             .order_by("diam")
             .min_monthly_budget_units("ipsum")
             .min_monthly_budget_nanos(-5)
             .min_monthly_budget_currency_code("et")
             .max_monthly_budget_units("duo")
             .max_monthly_budget_nanos(-32)
             .max_monthly_budget_currency_code("sea")
             .add_language_codes("Lorem")
             .add_industries("eos")
             .add_gps_motivations("erat")
             .company_name("sadipscing")
             .address("dolor")
             .doit();

Methods

impl<'a, C, A> CompanyListCall<'a, C, A> where C: BorrowMut<Client>, A: GetToken

fn doit(self) -> Result<(Response, ListCompaniesResponse)>

Perform the operation you have build so far.

fn website_url(self, new_value: &str) -> CompanyListCall<'a, C, A>

Website URL that will help to find a better matched company. .

Sets the website url query property to the given value.

fn view(self, new_value: &str) -> CompanyListCall<'a, C, A>

The view of the Company resource to be returned. This must not be COMPANY_VIEW_UNSPECIFIED.

Sets the view query property to the given value.

fn add_services(self, new_value: &str) -> CompanyListCall<'a, C, A>

List of services the company can help with.

Append the given value to the services query property. Each appended value will retain its original ordering and be '/'-separated in the URL's parameters.

fn request_metadata_user_overrides_user_id(self, new_value: &str) -> CompanyListCall<'a, C, A>

Logged-in user ID to impersonate instead of the user's ID.

Sets the request metadata.user overrides.user id query property to the given value.

fn request_metadata_user_overrides_ip_address(self, new_value: &str) -> CompanyListCall<'a, C, A>

IP address to use instead of the user's geo-located IP address.

Sets the request metadata.user overrides.ip address query property to the given value.

fn request_metadata_traffic_source_traffic_sub_id(self, new_value: &str) -> CompanyListCall<'a, C, A>

Second level identifier to indicate where the traffic comes from. An identifier has multiple letters created by a team which redirected the traffic to us.

Sets the request metadata.traffic source.traffic sub id query property to the given value.

fn request_metadata_traffic_source_traffic_source_id(self, new_value: &str) -> CompanyListCall<'a, C, A>

Identifier to indicate where the traffic comes from. An identifier has multiple letters created by a team which redirected the traffic to us.

Sets the request metadata.traffic source.traffic source id query property to the given value.

fn request_metadata_partners_session_id(self, new_value: &str) -> CompanyListCall<'a, C, A>

Google Partners session ID.

Sets the request metadata.partners session id query property to the given value.

fn request_metadata_locale(self, new_value: &str) -> CompanyListCall<'a, C, A>

Locale to use for the current request.

Sets the request metadata.locale query property to the given value.

fn add_request_metadata_experiment_ids(self, new_value: &str) -> CompanyListCall<'a, C, A>

Experiment IDs the current request belongs to.

Append the given value to the request metadata.experiment ids query property. Each appended value will retain its original ordering and be '/'-separated in the URL's parameters.

fn page_token(self, new_value: &str) -> CompanyListCall<'a, C, A>

A token identifying a page of results that the server returns. Typically, this is the value of ListCompaniesResponse.next_page_token returned from the previous call to ListCompanies.

Sets the page token query property to the given value.

fn page_size(self, new_value: i32) -> CompanyListCall<'a, C, A>

Requested page size. Server may return fewer companies than requested. If unspecified, server picks an appropriate default.

Sets the page size query property to the given value.

fn order_by(self, new_value: &str) -> CompanyListCall<'a, C, A>

How to order addresses within the returned companies. Currently, only address and address desc is supported which will sorted by closest to farthest in distance from given address and farthest to closest distance from given address respectively.

Sets the order by query property to the given value.

fn min_monthly_budget_units(self, new_value: &str) -> CompanyListCall<'a, C, A>

The whole units of the amount. For example if currencyCode is "USD", then 1 unit is one US dollar.

Sets the min monthly budget.units query property to the given value.

fn min_monthly_budget_nanos(self, new_value: i32) -> CompanyListCall<'a, C, A>

Number of nano (10-9) units of the amount. The value must be between -999,999,999 and +999,999,999 inclusive. If units is positive, nanos must be positive or zero. If units is zero, nanos can be positive, zero, or negative. If units is negative, nanos must be negative or zero. For example $-1.75 is represented as units=-1 and nanos=-750,000,000.

Sets the min monthly budget.nanos query property to the given value.

fn min_monthly_budget_currency_code(self, new_value: &str) -> CompanyListCall<'a, C, A>

The 3-letter currency code defined in ISO 4217.

Sets the min monthly budget.currency code query property to the given value.

fn max_monthly_budget_units(self, new_value: &str) -> CompanyListCall<'a, C, A>

The whole units of the amount. For example if currencyCode is "USD", then 1 unit is one US dollar.

Sets the max monthly budget.units query property to the given value.

fn max_monthly_budget_nanos(self, new_value: i32) -> CompanyListCall<'a, C, A>

Number of nano (10-9) units of the amount. The value must be between -999,999,999 and +999,999,999 inclusive. If units is positive, nanos must be positive or zero. If units is zero, nanos can be positive, zero, or negative. If units is negative, nanos must be negative or zero. For example $-1.75 is represented as units=-1 and nanos=-750,000,000.

Sets the max monthly budget.nanos query property to the given value.

fn max_monthly_budget_currency_code(self, new_value: &str) -> CompanyListCall<'a, C, A>

The 3-letter currency code defined in ISO 4217.

Sets the max monthly budget.currency code query property to the given value.

fn add_language_codes(self, new_value: &str) -> CompanyListCall<'a, C, A>

List of language codes that company can support. Only primary language subtags are accepted as defined by BCP 47 (IETF BCP 47, "Tags for Identifying Languages").

Append the given value to the language codes query property. Each appended value will retain its original ordering and be '/'-separated in the URL's parameters.

fn add_industries(self, new_value: &str) -> CompanyListCall<'a, C, A>

List of industries the company can help with.

Append the given value to the industries query property. Each appended value will retain its original ordering and be '/'-separated in the URL's parameters.

fn add_gps_motivations(self, new_value: &str) -> CompanyListCall<'a, C, A>

List of reasons for using Google Partner Search to get companies.

Append the given value to the gps motivations query property. Each appended value will retain its original ordering and be '/'-separated in the URL's parameters.

fn company_name(self, new_value: &str) -> CompanyListCall<'a, C, A>

Company name to search for.

Sets the company name query property to the given value.

fn address(self, new_value: &str) -> CompanyListCall<'a, C, A>

The address to use when searching for companies. If not given, the geo-located address of the request is used.

Sets the address query property to the given value.

fn delegate(self, new_value: &'a mut Delegate) -> CompanyListCall<'a, C, A>

The delegate implementation is consulted whenever there is an intermediate result, or if something goes wrong while executing the actual API request.

It should be used to handle progress information, and to implement a certain level of resilience.

Sets the delegate property to the given value.

fn param<T>(self, name: T, value: T) -> CompanyListCall<'a, C, A> where T: AsRef<str>

Set any additional parameter of the query string used in the request. It should be used to set parameters which are not yet available through their own setters.

Please note that this method must not be used to set any of the known paramters which have their own setter method. If done anyway, the request will fail.

Additional Parameters

bearer_token (query-string) - OAuth bearer token.

pp (query-boolean) - Pretty-print response.

prettyPrint (query-boolean) - Returns response with indentations and line breaks.

upload_protocol (query-string) - Upload protocol for media (e.g. "raw", "multipart").

access_token (query-string) - OAuth access token.

fields (query-string) - Selector specifying which fields to include in a partial response.

quotaUser (query-string) - Available to use for quota purposes for server-side applications. Can be any arbitrary string assigned to a user, but should not exceed 40 characters.

callback (query-string) - JSONP

oauth_token (query-string) - OAuth 2.0 token for the current user.

key (query-string) - API key. Your API key identifies your project and provides you with API access, quota, and reports. Required unless you provide an OAuth 2.0 token.

uploadType (query-string) - Legacy upload protocol for media (e.g. "media", "multipart").

alt (query-string) - Data format for response.

$.xgafv (query-string) - V1 error format.