Struct google_partners2::CompanyMethods

pub struct CompanyMethods<'a, C, A> where C: 'a, A: 'a {
    // some fields omitted
}

A builder providing access to all methods supported on company resources. It is not used directly, but through the Partners hub.

Example

Instantiate a resource builder

extern crate hyper;
extern crate yup_oauth2 as oauth2;
extern crate google_partners2 as partners2;
 
use std::default::Default;
use oauth2::{Authenticator, DefaultAuthenticatorDelegate, ApplicationSecret, MemoryStorage};
use partners2::Partners;
 
let secret: ApplicationSecret = Default::default();
let auth = Authenticator::new(&secret, DefaultAuthenticatorDelegate,
                              hyper::Client::new(),
                              <MemoryStorage as Default>::default(), None);
let mut hub = Partners::new(hyper::Client::new(), auth);
// Usually you wouldn't bind this to a variable, but keep calling *CallBuilders*
// like `get(...)`, `leads_create(...)` and `list(...)`
// to build up your call.
let rb = hub.companies();

Methods

impl<'a, C, A> CompanyMethods<'a, C, A>

fn get(&self, company_id: &str) -> CompanyGetCall<'a, C, A>

Create a builder to help you perform the following task:

Gets a company.

Arguments

companyId - The ID of the company to retrieve.

fn leads_create(&self, request: CreateLeadRequest, company_id: &str) -> CompanyLeadCreateCall<'a, C, A>

Create a builder to help you perform the following task:

Creates an advertiser lead for the given company ID.

Arguments

request - No description provided.

companyId - The ID of the company to contact.

fn list(&self) -> CompanyListCall<'a, C, A>

Create a builder to help you perform the following task:

Lists companies.