Struct google_partners2::Lead

pub struct Lead {
    pub gps_motivations: Option<Vec<String>>,
    pub family_name: Option<String>,
    pub min_monthly_budget: Option<Money>,
    pub comments: Option<String>,
    pub id: Option<String>,
    pub website_url: Option<String>,
    pub phone_number: Option<String>,
    pub given_name: Option<String>,
    pub type_: Option<String>,
    pub email: Option<String>,
}

A lead resource that represents an advertiser contact for a Company. These are usually generated via Google Partner Search (the advertiser portal).

This type is not used in any activity, and only used as part of another schema.

Fields

gps_motivations

List of reasons for using Google Partner Search and creating a lead.

family_name

Last name of lead source.

min_monthly_budget

The minimum monthly budget lead source is willing to spend.

comments

Comments lead source gave.

id

ID of the lead.

website_url

Website URL of lead source.

phone_number

Phone number of lead source.

given_name

First name of lead source.

type_

Type of lead.

email

Email address of lead source.