Struct google_partners2::ListCompaniesResponse

pub struct ListCompaniesResponse {
    pub next_page_token: Option<String>,
    pub companies: Option<Vec<Company>>,
    pub response_metadata: Option<ResponseMetadata>,
}

Response message for ListCompanies.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

list companies (response)

Fields

next_page_token

A token to retrieve next page of results. Pass this value in the ListCompaniesRequest.page_token field in the subsequent call to ListCompanies to retrieve the next page of results.

companies

The list of companies.

response_metadata

Current response metadata.