Struct google_partners2::UserStateListCall

pub struct UserStateListCall<'a, C, A> where C: 'a, A: 'a {
    // some fields omitted
}

Lists states for current user.

A builder for the list method supported by a userState resource. It is not used directly, but through a UserStateMethods instance.

Example

Instantiate a resource method builder

 
// You can configure optional parameters by calling the respective setters at will, and
// execute the final call using `doit()`.
// Values shown here are possibly random and not representative !
let result = hub.user_states().list()
             .request_metadata_user_overrides_user_id("eirmod")
             .request_metadata_user_overrides_ip_address("elitr")
             .request_metadata_traffic_source_traffic_sub_id("amet")
             .request_metadata_traffic_source_traffic_source_id("no")
             .request_metadata_partners_session_id("labore")
             .request_metadata_locale("eirmod")
             .add_request_metadata_experiment_ids("dolore")
             .doit();

Methods

impl<'a, C, A> UserStateListCall<'a, C, A> where C: BorrowMut<Client>, A: GetToken

fn doit(self) -> Result<(Response, ListUserStatesResponse)>

Perform the operation you have build so far.

fn request_metadata_user_overrides_user_id(self, new_value: &str) -> UserStateListCall<'a, C, A>

Logged-in user ID to impersonate instead of the user's ID.

Sets the request metadata.user overrides.user id query property to the given value.

fn request_metadata_user_overrides_ip_address(self, new_value: &str) -> UserStateListCall<'a, C, A>

IP address to use instead of the user's geo-located IP address.

Sets the request metadata.user overrides.ip address query property to the given value.

fn request_metadata_traffic_source_traffic_sub_id(self, new_value: &str) -> UserStateListCall<'a, C, A>

Second level identifier to indicate where the traffic comes from. An identifier has multiple letters created by a team which redirected the traffic to us.

Sets the request metadata.traffic source.traffic sub id query property to the given value.

fn request_metadata_traffic_source_traffic_source_id(self, new_value: &str) -> UserStateListCall<'a, C, A>

Identifier to indicate where the traffic comes from. An identifier has multiple letters created by a team which redirected the traffic to us.

Sets the request metadata.traffic source.traffic source id query property to the given value.

fn request_metadata_partners_session_id(self, new_value: &str) -> UserStateListCall<'a, C, A>

Google Partners session ID.

Sets the request metadata.partners session id query property to the given value.

fn request_metadata_locale(self, new_value: &str) -> UserStateListCall<'a, C, A>

Locale to use for the current request.

Sets the request metadata.locale query property to the given value.

fn add_request_metadata_experiment_ids(self, new_value: &str) -> UserStateListCall<'a, C, A>

Experiment IDs the current request belongs to.

Append the given value to the request metadata.experiment ids query property. Each appended value will retain its original ordering and be '/'-separated in the URL's parameters.

fn delegate(self, new_value: &'a mut Delegate) -> UserStateListCall<'a, C, A>

The delegate implementation is consulted whenever there is an intermediate result, or if something goes wrong while executing the actual API request.

It should be used to handle progress information, and to implement a certain level of resilience.

Sets the delegate property to the given value.

fn param<T>(self, name: T, value: T) -> UserStateListCall<'a, C, A> where T: AsRef<str>

Set any additional parameter of the query string used in the request. It should be used to set parameters which are not yet available through their own setters.

Please note that this method must not be used to set any of the known paramters which have their own setter method. If done anyway, the request will fail.

Additional Parameters

bearer_token (query-string) - OAuth bearer token.

pp (query-boolean) - Pretty-print response.

prettyPrint (query-boolean) - Returns response with indentations and line breaks.

upload_protocol (query-string) - Upload protocol for media (e.g. "raw", "multipart").

access_token (query-string) - OAuth access token.

fields (query-string) - Selector specifying which fields to include in a partial response.

quotaUser (query-string) - Available to use for quota purposes for server-side applications. Can be any arbitrary string assigned to a user, but should not exceed 40 characters.

callback (query-string) - JSONP

oauth_token (query-string) - OAuth 2.0 token for the current user.

key (query-string) - API key. Your API key identifies your project and provides you with API access, quota, and reports. Required unless you provide an OAuth 2.0 token.

uploadType (query-string) - Legacy upload protocol for media (e.g. "media", "multipart").

alt (query-string) - Data format for response.

$.xgafv (query-string) - V1 error format.